# courses/backend/node/nodejs-week3/openapi.yml
openapi: 3.0.0
info:
  title: Snippets API
  version: 1.0.0
  description: API for managing code snippets

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /api/snippets:
    get:
      summary: Get all snippets
      responses:
        "200":
          description: List of all snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - contents
              properties:
                title:
                  type: string
                  description: Title of the snippet
                contents:
                  type: string
                  description: Code contents of the snippet
      responses:
        "201":
          description: Snippet created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/snippets/unsafe:
    get:
      summary: Get all snippets with safe sorting
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            description: Sort by column and direction, e.g., 'title DESC'
          example: "title DESC"
      responses:
        "200":
          description: List of sorted snippets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Snippet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/snippets/public:
    get:
      summary: Get public snippets
      responses:
        "200":
          description: List of public snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/snippets/tag/{tag}:
    get:
      summary: Get snippets by tag
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
          description: Tag name
      responses:
        "200":
          description: List of snippets with the tag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/snippets/{id}:
    get:
      summary: Get snippet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Snippet ID
      responses:
        "200":
          description: Snippet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: Update snippet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Snippet ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the snippet
                contents:
                  type: string
                  description: Code contents of the snippet
      responses:
        "200":
          description: Snippet updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Snippet updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Delete snippet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Snippet ID
      responses:
        "200":
          description: Snippet deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Snippet deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    Snippet:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        title:
          type: string
          description: Title of the snippet
        contents:
          type: string
          description: Code contents
        user_id:
          type: integer
          description: User ID
          default: 1
        is_private:
          type: boolean
          description: Whether the snippet is private
          default: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - title
        - contents

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"